Odradjene transformacije (ne idealno)

+++ Cronus/src/engine/util/WindowCallbackObserver.h
@@ -7,6 +7,6 @@ class WindowCallbackObserver {
 public:
 	virtual void KeyInputUpdate(int key) = 0;
 	virtual void FrameBufferSizeUpdate(int width, int height) = 0;
-	virtual void MouseInputUpdate(double xPos, double yPos) = 0;
-	virtual void ScrollInputUpdate(double xOffset, double yOffset) = 0;
+	virtual void MouseInputUpdate(float xPos, float yPos) = 0;
+	virtual void ScrollInputUpdate(float xOffset, float yOffset) = 0;
 };
